requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.35
pdfplumber==0.11.4
python-docx==1.1.2
google-genai==1.70.0
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
requests==2.32.3
reportlab==4.2.2
aiofiles==24.1.0
websockets>=13.0
